+++ warp/_src/context.py
@@ -6260,7 +6260,6 @@ class Runtime:
                 ctypes.c_double,
                 ctypes.c_void_p,
                 ctypes.c_void_p,
-                ctypes.c_void_p,
             ]
             self.core.wp_hash_grid_reserve_host.argtypes = [ctypes.c_uint64, ctypes.c_int, ctypes.c_int]
 
@@ -6277,7 +6276,6 @@ class Runtime:
                 ctypes.c_uint64,
                 ctypes.c_int,
                 ctypes.c_double,
-                ctypes.c_void_p,
                 ctypes.c_void_p,
                 ctypes.c_void_p,
             ]